refactor: split main.ts into session manager + IPC handler modules (#7b of 2)
Closes the carve-out started in #42. main.ts is now a bootstrap shim:
window creation, MCP server stand-up (with the open_session
serialization queue), and IPC handler registration. Session lifecycle
and IPC handler bodies live in their own modules.

main.ts: 831 → 295 LOC (-536, -65%).

New modules:
- session-manager.ts (401 LOC) — Session type, sessions Map,
  statusEmitted Set, mainWindow ref, setStatus, shouldEmitStatus,
  findSessionByIdOrPrefix, persistSessions, deleteSession,
  createSessionInternal, killAllSessions. The mainWindow ref is
  injected via setMainWindow at startup so the module doesn't
  depend on main.ts.
- ipc-session.ts — registerSessionHandlers(): session:create /
  input / resize / close / list / rename + the shell-PTY variants.
- ipc-app.ts — registerAppHandlers({ config }): vscode:open,
  dialog:pickFolder, app:home, config:get, config:path,
  clipboard:read, clipboard:write, open-external, window:* . Has
  its own setMainWindow injection.
- ipc-discovery.ts — registerDiscoveryHandlers(): agents:list/open,
  skills:list/open with path-traversal guards.
- pty-resize.ts — resizePty (moved from main.ts).

Tests reorganized:
- Deleted electron/main.test.ts.
- buildClaudeArgs cases moved to claude-command.test.ts; the
  re-export from main.ts is gone.
- resizePty cases moved to pty-resize.test.ts.
- shouldEmitStatus cases moved to session-manager.test.ts (with
  electron + node-pty mocks so the module loads under Vitest).

Notable cleanups:
- Drop the mcpConfigPath module-level variable from main.ts. It was
  always equal to getMcpConfigPath() from config.ts, so
  session-manager calls the getter directly and main.ts just calls
  writeMcpConfigFile() for its file-creation side effect.
- Drop the buildClaudeArgs re-export from main.ts (was a transitional
  step from #7a).

Verified: typecheck (both tsconfigs), npm test (98/98 — same count,
just reorganized), npm run build.
